requirements.txt
streamlit
opencv-python-headless
pytesseract
gTTS
pillow
numpy
